requirements.txt
requests
Pillow
beatifulsoup4
lxml